Revert "indicesをu32からusizeに変更"
This reverts commit 074b45527ecb8d2a2e87b032c4dbe32237dcc0a2.

+++ src/tensor/mod.rs
@@ -731,7 +731,7 @@ impl TensorOps for Tensor {
         ))
     }
 
-    fn rows_slice(&self, indices: &[usize]) -> Result<Self> {
+    fn rows_slice(&self, indices: &[u32]) -> Result<Self> {
         let result_shape = Shape::new(vec![indices.len(), self.shape.dims()[1]]).unwrap();
         for backend in &BACKENDS[0..] {
             match backend.rows_slice(&self.storage, &self.shape, &indices) {
